small fixes

+++ cli/cli_app.go
@@ -377,7 +377,7 @@ func RunTerragrunt(terragruntOptions *options.TerragruntOptions) error {
 
 	// replace default sts duration if set in config
 	defaultIamAssumeRoleDuration := 3600 // Default duration for IAM Assume role session is 3600s (1hr)
-	if terragruntOptions.IamAssumeRoleDuration == defaultIamAssumeRoleDuration && terragruntConfig.IamAssumeRoleDuration != nil {
+	if terragruntOptions.IamAssumeRoleDuration == int64(defaultIamAssumeRoleDuration) && terragruntConfig.IamAssumeRoleDuration != nil {
 		terragruntOptions.IamAssumeRoleDuration = *terragruntConfig.IamAssumeRoleDuration
 	}
